+++ lib/mongoid/history/hooks.rb
@@ -1,0 +1,13 @@
+module Mongoid
+  module History
+    module Hooks
+      extend ActiveSupport::Concern
+
+      included do
+        before_action do |controller|
+          Thread.current[:mongoid_history_controller] = controller
+        end
+      end
+    end
+  end
+end