+++ lib/mongoid/history/tracker.rb
@@ -14,7 +14,12 @@ module Mongoid
         field :version,                 type: Integer
         field :action,                  type: String
         field :scope,                   type: String
-        belongs_to :modifier, class_name: Mongoid::History.modifier_class_name
+        if respond_to? :t_belongs_to
+          # Tenacity support https://github.com/jwood/tenacity
+          t_belongs_to :modifier, class_name: Mongoid::History.modifier_class_name
+        else
+          belongs_to :modifier, class_name: Mongoid::History.modifier_class_name
+        end
 
         index(scope: 1)
         index(association_chain: 1)